toswe/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.2
cryptography==45.0.5
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
jwt==1.4.0
msgpack==1.1.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
